train_requirements.txt
torch
numpy
pyyaml
tqdm
tensorboard